Added dirty get agents tasks for enabling further PoC tests

+++ packages/cw-croncat-core/src/msg.rs
@@ -145,7 +145,7 @@ pub struct GetAgentIdsResponse {
 }
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
-pub struct GetAgentTasksResponse(pub u64, pub u128);
+pub struct GetAgentTasksResponse(pub u64, pub u64);
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
 pub struct TaskRequest {
